+++ generators/app/index.js
@@ -68,10 +68,7 @@ class ReactNativeBootstrap extends Generator {
       this.features = features;
       this.features.landscape = landscape;
       this.features.hasFirebase =
-        features.crashlytics ||
-        features.firebaseanalytics ||
-        features.pushnotifications ||
-        features.firebaseperformance;
+        features.firebasecrashlytics || features.firebaseanalytics || features.firebaseperformance;
       return this.prompt([
         {
           type: 'input',