+++ generators/app/tasks/appSetup/coreFiles/androidProjectSetup.js
@@ -57,14 +57,18 @@ function updateAppProguardRules() {
   const proguardRulesContent = this.fs.read(`${this.projectName}/android/app/proguard-rules.pro`);
   const updatedProguardRulesContent = proguardRulesContent.replace(
     '# Add any project specific keep options here:',
-    `# Add any project specific keep options here:\n# Hermes\n-keep class com.facebook.hermes.unicode.** { *; }\n\n# react-native-config\n-keep class com.${this.projectName.toLowerCase()}.BuildConfig { *; }`
+    `# Add any project specific keep options here:\n# Hermes\n-keep class com.facebook.hermes.unicode.** { *; }\n-keep class com.facebook.jni.** { *; }\n\n# react-native-config\n-keep class com.${this.projectName.toLowerCase()}.BuildConfig { *; }`
   );
   this.fs.write(`${this.projectName}/android/app/proguard-rules.pro`, updatedProguardRulesContent);
 }
 
-function disableR8ForReleases() {
+function updateGradleProperties() {
   const gradleProperties = this.fs.read(`${this.projectName}/android/gradle.properties`);
-  const updatedGradleProperties = gradleProperties.replace(
+  let updatedGradleProperties = gradleProperties.replace(
+    '# org.gradle.jvmargs=-Xmx2048m -XX:MaxPermSize=512m -XX:+HeapDumpOnOutOfMemoryError -Dfile.encoding=UTF-8',
+    'org.gradle.jvmargs=-Xmx2048m -XX:MaxPermSize=512m -XX:+HeapDumpOnOutOfMemoryError -Dfile.encoding=UTF-8'
+  );
+  updatedGradleProperties = updatedGradleProperties.replace(
     'android.enableJetifier=true',
     'android.enableJetifier=true\nandroid.enableR8=false'
   );
@@ -75,5 +79,5 @@ module.exports = function androidProjectSetup() {
   updateAppBuildGradle.bind(this)();
   addRNGestureHandlerConfig.bind(this)();
   updateAppProguardRules.bind(this)();
-  disableR8ForReleases.bind(this)();
+  updateGradleProperties.bind(this)();
 };