+++ generators/app/templates/src/constants/serializers.ts
@@ -1,0 +1,4 @@
+import { CamelcaseSerializer, SnakecaseSerializer } from 'cerealizr';
+
+export const snakeCaseSerializer = new SnakecaseSerializer();
+export const camelCaseSerializer = new CamelcaseSerializer();